requirements.txt
aiohttp==3.6.2
python-dotenv==0.13.0